requirements.txt
# HTTP Client
requests>=2.31.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Date/Time Parsing
python-dateutil>=2.8.0

# JSON Schema Validation
jsonschema>=4.20.0

# Configuration
pyyaml>=6.0.1

# Fuzzy String Matching
rapidfuzz>=3.6.0

# Development/Testing
pytest>=8.0.0
pytest-cov>=4.1.0
